requirements.txt
annotated-types==0.7.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
docstring_parser==0.16
Flask==3.0.3
google-ai-generativelanguage==0.4.0
google-api-core==2.24.2
google-auth==2.39.0
google-cloud-aiplatform==1.71.1
google-cloud-bigquery==3.31.0
google-cloud-core==2.4.3
google-cloud-language==2.13.4
google-cloud-resource-manager==1.14.2
google-cloud-speech==2.26.0
google-cloud-storage==2.16.0
google-cloud-texttospeech==2.17.2
google-crc32c==1.7.1
google-genai==0.1.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.72.0rc1
grpcio-status==1.63.0rc1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.4
packaging==24.2
pillow==11.1.0
proto-plus==1.26.1
protobuf==4.25.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic_core==2.33.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
requests==2.31.0
rsa==4.9
setuptools==78.1.0
shapely==2.1.0
six==1.17.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
vertexai==1.71.1
websockets==14.2
Werkzeug==3.1.3
wheel==0.45.1
